1. answer
1. typed the code.

Made the cose a little more efficient be defining i in a for loop (refer to line 18)
2. tested the code with the python shell and the program works.
3.Divided the code into an if else statement to determine a valid guess vs an invalid guess.
Had to tab in line 30 to make it part of the letter in A condition.
4. made a counter variable called wrong to track the number of wrong guesses.
In the condition when "wrong" reaches 6, the program will end and print "GAME OVER!!" (refer to lines 42 to 44)
5.
Imported the random class, then made a list called "words", which contains four words: hangman, great, computer, and science. rand will pick a word from list "words" (Refer to lines 2 to 18)